+++ src/main/java/com/github/jnhyperion/hyperrobotframeworkplugin/psi/element/DefinedParameter.java
@@ -16,6 +16,8 @@ public interface DefinedParameter extends LookupElementMarker {
    @Nullable
    String getDefaultValue();
 
+   boolean isKeywordContainer();
+
    @NotNull
    PsiElement reference();
 }